requirements.txt
fastapi==0.116.1
uvicorn==0.22.0
databases[postgresql]==0.11.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.7
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
